Day7

+++ src/day6.rs
@@ -5,18 +5,18 @@ fn num_lanternfish(filename: &str, days: usize) -> Result<usize> {
     let numbers = parse_numbers_comma(filename)?;
     let mut counts = [0; 9];
     for n in numbers {
-        counts[n] += 1;
+        counts[n as usize] += 1;
     }
 
     for _day in 0..days {
         let zero_count = counts[0];
         for i in 0..8 {
-            counts[i] = counts[i+1]
+            counts[i] = counts[i + 1]
         }
         counts[8] = zero_count;
         counts[6] += zero_count;
     }
-    
+
     Ok(counts.iter().sum())
 }
 
@@ -27,7 +27,6 @@ fn q1(filename: &str) -> Result<usize> {
 fn q2(filename: &str) -> Result<usize> {
     num_lanternfish(filename, 256)
 }
-
 
 #[cfg(test)]
 mod tests {
